requirements.txt
Flask==0.11.*
requests
Werkzeug==0.11.*
privex-exchange>=0.1.0
privex-helpers>=2.13.0
nest_asyncio
jupyter
python-dotenv
gunicorn

